refactor: extract shared exec file I/O; dedupe test runtime boilerplate; drop dead code
